refactor: 댓글 목 api 한 파일에 관리 및 댓글 post 인풋에서만 처리하게 수정
